+++ hooks/auth/useAdminAuth.js
@@ -1,0 +1,23 @@
+import { useEffect, useState } from "react";
+import { onAuthStateChanged } from "firebase/auth";
+import { auth } from "@/lib/firebase/client";
+
+export function useAdminAuth() {
+  const [authReady, setAuthReady] = useState(false);
+  const [authUser, setAuthUser] = useState(null);
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      setAuthUser(user);
+      setAuthReady(true);
+    });
+    return () => unsubscribe();
+  }, []);
+
+  const getToken = async () => {
+    if (!authUser) return null;
+    return await authUser.getIdToken();
+  };
+
+  return { authReady, authUser, getToken };
+}
